Rename profile::base to nest
When I started out, my original intention was to support many operating
systems, each extending off of a base profile, but I quickly gave up on
that idea--one OS is enough for me to manage--and Funtoo has proven
itself to be more than capable and flexible for anything I throw at it:
desktops, servers, and even ARM dev boards.  So this rename reflects my
committment to Funtoo for now.