Add hosts entry for unifi.home
The USG creates an CNAME unifi.home=unifi, and an A record
unifi=172.22.1.12, but systemd-resolved reserves dotless records for
LLMNR and I can't find a workaround.  This "solution" is the lowest
touch fix and shifts the responsibility of the workaround to my hosts
rather than other hacks we can do in UniFi.