timesyncd: Use NIST time servers by IP
Workaround https://github.com/systemd/systemd/issues/5873 by configuring
timesyncd with a list of reliable, public time servers by IP address.

NIST provides my country's official time and my watches sync to it by
WWV so this "workaround" appeals to me in more ways than one, even
though I only really need accuracy down to a second.

Initially I wanted to append the list of IPs behind the service name
time.nist.gov so that it would be preferred, but timesyncd seems to
append the DNS lookup results to the list of IPs, making it both
redundant and not likely to be reached.