firefox: Tweak env vars for 83.0
MOZ_USE_XINPUT2 is stable to me and can apply to Xorg and Xwayland and
it doesn't affect Wayland sessions, so enable it by default.

MOZ_X11_EGL can be used to enable VAAPI acceleration and seems to work
well on Xorg, but it causes artifacts under Xwayland--just enable on
Xorg.

Also, WebRender seems to be working pretty well on Panfrost, so now that
I've gotten rid of Nvidia hardware, enable WebRender by default, except
on raspberrypi, which is still needs more work.