firefox: Move all-nest.js out of wrapper
I forgot that javascript was a programming language.  Moving these
preferences out of the wrapper ensures the settings work for multiple
users (not that that is my use case) and provides me a better place to
define a new setting, network.trr.mode, for enabling DNS over TLS to
workaround a systemd-resolved bug.