gitlab-runner: fuse-overlayfs needs privileged container
I can't find the magic combination of capabilities needed to make
fuse-overlayfs, and it's not impossible that Fedora have patched or more
updated version of whatever is necessary to make this work.  Ultimately,
this runner will only be running my custom image which I already run as
root on the host, so privileged doesn't really matter here.