bitwarden: Remove dns parameter
This was meant to let the container resolve web.nest, but it seems to
work just fine without.  The dns parameter is absolutely required for
the linuxserver containers, which reliably fail to look up nest hosts
without an explicit configuration to use nest DNS.  I think that has
more to do with the way those containers are made.