gitlab-runner: Kill /dev/fuse
Let containers-in-containers (buildah) use VFS, not fuse-overlayfs,
which is still buggy.  I tracked a problem compiling ncurses down to
fuse-overlayfs.  It seems like fuse-overlayfs has problems with quick
create, write, read, delete in succession to the same file, and it is
resulting in a race condition that causes an ncurses build script to
fail.  VFS doesn't support layered images without copying, but I don't
do much work with layers, so I'll take the stability and relative
simplicity of the VFS system.