xmonad: Bootstrap GHC on non-AMD64 architectures
The Gentoo Haskell project only provides bootstrapping binaries for
AMD64, with the exception of version 8.6.5, which has binaries for most
architectures.  Install that one first, then bootstrap an updated GHC
from there.  Do this in the xmonad class because that's the only thing
that requires and pulls in GHC.  If that changes, move GHC to its own
class.