zfs: Turn Dracut prompt tweak into patch
This is a code change I think should be upstreamed for consistency with
the cryptsetup module and to clarify terms ("passphrase", encryption
root).  Tired of seeing the diff come up in Puppet after upgrades for
something that is not fixing a configuration.

Also, remove old script that has long since been removed from service.