u-boot: Disable OHCI instead of preboot cmd
The preboot command containing "usb start" is also in the distro boot
command so disabling the preboot command doesn't help.  Disabling
OHCI_GENERIC worked except for when plugged into the Pine64 dock.  That
must have an OHCI interface because it hangs the system at boot until
the OHCI_HCD feature is disabled.