vmware: Force smallest console font size
VMware Fusion *will not* allow for legacy video memory > 4MB so the
console framebuffer resolution can't be increased beyond 800x600. There
is a setting svga.guestBackedPrimaryAware that can be set to FALSE to
workaround the problem, but VMware resets it upon any settings changes
or even a clean shutdown. Don't fight it for a console I rarely use and
just use a smaller font.