k8s: Revert to default 'LimitedSwap' setting
The UnlimitedSwap value is broken with CRI-O 1.23. I think it worked
under 1.22, but the whole swap support is not considered stable yet, so
don't worry about it. The exact error that popped up with the software
versions is:

create failed because memory swap limit (-1) cannot be lower than memory limit