k8s: Fix firewall rule for Calico pod network
Pod network packets are handled by forwarding filters. Firewalld
defaults to allow forwarding for known networks, but the catch is that
the network must be known to the default zone. So, counter-intuitively,
we are adding this pod network range to the "drop" zone, simply because
that is the default zone. Use resource tagging to make that more clear.