dracut: Add delay to ensure all devices are ready before ZFS
Upon moving to USB storage for Raspberry Pi, I found the device was not
initialized by the kernel before dracut started the
zfs-import-cache/scan service, resulting in the pool not being imported.
The kernel initializes devices asynchronously and, short of knowing
which device you're looking for and waiting for it specifically, we need
to add a generic delay for devices to become available.

Never happy adding hard-coded delays. Hopefully five seconds does the
trick.