kernel: Fix exec return code and install bc
The main kernel build status was not returned due to the presence of
`tee` at the end of the pipeline. Run in a shell with `pipefail` set.

Also, `bc` seems to be something that the kernel source packages used to
pull in. Now that the sources are coming from Git, we need to make sure
all the dependencies are installed.