rockpro64: Use older U-Boot
U-Boot 2023.04-rc2 hangs at the SPL while trying to load the main image.
Maybe something to do with a change in how the image is packed or
compiled. Needs further investigation, but for now, just use an earlier
version that seems to work.

Also, reduce the RK3399 baud rate to a more common and stable 115,200.