k8s: Define backup and sidecar once for all apps
Refactor the way resources and patches are looked up to allow the
definitions for the Nest backup and sidecars to be moved to the default
Hiera level. Allow for more expressive configurations and easier
interactions between patches at different levels of Hiera.

Also, re-add pdk.yaml lol. It appears after a `pdk validate`.