build: Major refactor of class declaration
Use Hiera to specify build targets and the classes and parameters to
declare. Decouple the whole concept of build stages, roles, and tools
from the code itself, especially for platforms like Windows (or future
non-Gentoo bases) where it doesn't make as much sense. Let sensitive
data be defined directly in the stages where it is used to eliminate
potential leaks and a lot of conditional code that checks the `build`
fact. Flatten the manifests directory structure to make the workstation,
now "gui", classes more accessible. Correct behavior is confirmed with
rspec.