milkv-pioneer: Switch back to normal defconfig
The SMP bring-up problem was due to my ZSBL. Still figuring it out.
Until then, get a good kernel going and a "working" U-Boot. U-Boot seems
to modify the memory maps such that Linux only sees 16 GB, but
everything seems to work under u-root, which is probably the next step.