kernel: Rework zfs-kmod and module-rebuild
The `noop` metaparameter doesn't seem to be propagating to
`nest::lib::package` resources without also passing it as a regular
parameter, which puppet-syntax flags as a warning. Also reorder the
resources to ensure module-rebuild doesn't build an outdated version of
zfs-kmod. This might cause zfs-kmod to be built twice in the worst
case--to upgrade and as part of module-rebuild.