pdk: Fix module initialization
Somewhere along the line, the bundle update that happens during module
creation started failing on certain platforms:

> pdk (INFO): Creating new module: build
> pdk (INFO): Using the default template-url and template-ref.
> [✔] Resolving default Gemfile dependencies.
> pdk (FATAL): Writing lockfile to /module/Gemfile.lock
>
> Could not find gem 'voxpupuli-puppet-lint-plugins (~> 5.0)' in locally installed gems.
>
> pdk (FATAL): Unable to resolve Gemfile dependencies.

This is a gem referenced in the module's Gemfile, but not pulled in by
the pdk gem, and since the bundle update has `local: true`, it fails.
Just let bundler install missing gems. It's going into a container.