ceph: Monitor and report cluster utilization with prometheus
Compared to the broken checkpoint in the last commit, this one realizes
that the Prometheus instance needs to be in the same namespace as the
app, the rook-ceph-operator monitoring config was incorrect, and the
prometheus-operator needs an explicit ServiceMonitor selector config.
Aggressively use affinity to force all this stuff onto the dedicated
storage nodes.