k8s: Don't use SSH agent for bolt
The SSH agent forwarding doesn't work when using proxy jump with Bolt.
Map the ssh key into the sidecar for other SSH activities like rsync. If
not using SSH agent, provide explicit Bolt configuration to find private
key. Interact with k8s services as 'james'. That is the way the sidecar
was intended to be used.