eyrie: Fix restore with jump box
The inventory data for k8s services is critical now that it contains the
service and proxyjump info. Unfortunately, we can't refresh the Bolt
inventory while it's running, at least not easily, so modify the plans
to ensure the restore works on the second run. Add new `init` parameter
to deploy plans to avoid failure on the first deployment.

> bolt plan run nest::eyrie::deploy_portfolio init=true
> bolt plan run nest::eyrie::deploy_portfolio restore=true --stream
> bolt plan run nest::eyrie::deploy_portfolio # enables backups