k8s: Deploy test jump plus load balancing fixes
Advertise test IP pool on workstations only. This keeps traffic as
direct as possible. In the process, refactor node role labels to use the
(internal but ok) node-role.kubernetes.io special name. The only effect
of this is to show the role in `kubectl get nodes`, which is nice. We
can also assign multiple roles to a node with this labeling scheme.