mariadb: Switch back to ceph-block
Restore the original vaultwarden check intervals. The poor performance
was caused by latency on the OSDs which I tracked down to the SSDs
themselves. They're just not up to the heavy synchronous write load of
all the OSDs. Since this is just a test node and I don't want to buy new
drives, I just issued a `zfs set sync=disabled data` and the latency was
eliminated immediately. This could also be solved with a SLOG, which
moves the ZIL off the data SSDs to somewhere else for synchronous
writes, which ZFS then manages asynchronously back to the SSDs.
Unfortunately, I don't have the drives or the space.