ceph: Disable 'sync' on test SSDs for performance
Accept the possibility of up to 5 seconds of data loss in test for about
a 100x speedup on writing to the SSD pool in test. This is a situation
where we have three OSDs on one zpool (for testing!) and it can't handle
the 3x replicated sync requests from Ceph. Disabling sync should not
risk corruption because both Ceph and ZFS are transactionally safe.