k8s: Restore with init container
Rather than wait for a service to come up with default data and then try
to restore it, just restore it ahead of time. To avoid interfering with
pod launch almost every launch after the first, we'll remove the init
container from the deployment spec. This also avoids problems with the
registry not being availble, and lets us use ImagePullPolicy: Always, to
ensure we get the latest puppet code for the initialization.