ceph: Refine objectstore config and fix RGW dashboard
With ingress in place, we can move the cluster service port 8080 and set
`rgw_dns_name` in the deploy plan (see #66). Convert the plan to Puppet
language due to the new conditional complexity. Still needs to be done
for prod, but that is pending the deployment of an nginx-workstation
ingress controller such that all object traffic stays on
high-performance nodes.