pdk: Map eyaml config into container
Due to the Puppet and Bolt hiera consolidation and the way rspec-puppet
sets up fixtures, Bolt's `.modules` and the `data/private` dir are
visible to the spec test. This doesn't affect the CI pipelines because
they have a fresh clone and don't install Bolt modules before the test.
Revisit if this project gets refactored into separate module and control
repo.